[Refactor]: Remove utility functions from views into a seperate file
