+++ restapi/utils.py
@@ -1,0 +1,107 @@
+from datetime import datetime
+import urllib.request
+
+
+def normalize(expense) -> list:
+    '''Normalize the balances for a user'''
+    user_balances = expense.users.all()
+    dues = {}
+    for user_balance in user_balances:
+        dues[user_balance.user] = dues.get(user_balance.user, 0) + user_balance.amount_lent \
+            - user_balance.amount_owed
+    dues = [(k, v) for k, v in sorted(dues.items(), key=lambda item: item[1])]
+    start = 0
+    end = len(dues) - 1
+    balances = []
+    while start < end:
+        amount = min(abs(dues[start][1]), abs(dues[end][1]))
+        user_balance = {"from_user": dues[start][0].id,
+                        "to_user": dues[end][0].id, "amount": amount}
+        balances.append(user_balance)
+        dues[start] = (dues[start][0], dues[start][1] + amount)
+        dues[end] = (dues[end][0], dues[end][1] - amount)
+        if dues[start][1] == 0:
+            start += 1
+        else:
+            end -= 1
+    return balances
+
+
+def sort_by_time_stamp(logs) -> list:
+    '''Sort logs by Time Stamp'''
+    data = []
+    for log in logs:
+        data.append(log.split(" "))
+    data = sorted(data, key=lambda elem: elem[1])
+    return data
+
+
+def response_format(raw_data) -> list:
+    '''Format a response'''
+    response = []
+    for timestamp, data in raw_data.items():
+        entry = {'timestamp': timestamp}
+        logs = []
+        data = {k: data[k] for k in sorted(data.keys())}
+        for exception, count in data.items():
+            logs.append({'exception': exception, 'count': count})
+        entry['logs'] = logs
+        response.append(entry)
+    return response
+
+
+def aggregate(cleaned_logs) -> dict:
+    '''Aggregate cleaned logs'''
+    data = {}
+    for log in cleaned_logs:
+        [key, text] = log
+        value = data.get(key, {})
+        value[text] = value.get(text, 0)+1
+        data[key] = value
+    return data
+
+
+def transform(logs) -> list:
+    '''Clean up logs'''
+    result = []
+    for log in logs:
+        [_, timestamp, text] = log
+        text = text.rstrip()
+        timestamp = datetime.utcfromtimestamp(int(int(timestamp)/1000))
+        hours, minutes = timestamp.hour, timestamp.minute
+        key = ''
+
+        if minutes >= 45:
+            if hours == 23:
+                key = "{:02d}:45-00:00".format(hours)
+            else:
+                key = "{:02d}:45-{:02d}:00".format(hours, hours+1)
+        elif minutes >= 30:
+            key = "{:02d}:30-{:02d}:45".format(hours, hours)
+        elif minutes >= 15:
+            key = "{:02d}:15-{:02d}:30".format(hours, hours)
+        else:
+            key = "{:02d}:00-{:02d}:15".format(hours, hours)
+
+        result.append([key, text])
+        print(key)
+
+    return result
+
+
+def reader(url, timeout):
+    '''Read a file and return it's data'''
+    with urllib.request.urlopen(url, timeout=timeout) as conn:
+        # TODO: Check if this closes the open file
+        return conn.read()
+
+
+def multiThreadedReader(urls, num_threads) -> list:
+    """Read multiple files through HTTP"""
+    result = []
+    for url in urls:
+        data = reader(url, 60)
+        data = data.decode('utf-8')
+        result.extend(data.split("\n"))
+    result = sorted(result, key=lambda elem: elem[1])
+    return result
